+++ websiteapp/const.py
@@ -9,6 +9,8 @@
 # @license   https://www.opensource.org/licenses/mit-license.php MIT
 # @link      https://github.com/MarcinOrlowski/website-as-app
 #
+# @file      websiteapp/const.py
+#
 ##################################################################################
 """
 
@@ -18,7 +20,7 @@ from typing import List
 
 class Version(Enum):
     MAJOR = 1
-    MINOR = 3
+    MINOR = 4
     PATCH = 0
 
     @classmethod