+++ CHANGES.md
@@ -4,8 +4,12 @@
 
 ---
 
-* @dev
+* 1.4.0 (2024-11-12)
   * Added screenshot and simplified landing page.
+  * Added custom context menu for the embedded browser.
+  * Added `Copy URL` item to the context menu.
+  * Added `Open URL from Clipboard` item to the context menu.
+  * Added option to disable custom WebEngine used by the app (if needed).
 
 * 1.3.1 (2024-10-28)
   * Artificial release to fix isses with PyPy release.